pyproject.toml
[tool.poetry]
name = "url-shortener"
version = "0.1.0"
description = ""
authors = ["Albert <albertamoah2000@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.6"
uvicorn = {extras = ["standard"], version = "^0.34.0"}
redis = "^5.2.1"
python-dotenv = "^1.0.1"
fastapi-limiter = "^0.1.6"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
